feat: new names file structure
