+++ src/index.ts
@@ -1,15 +1,15 @@
 import express from 'express'
 import cors from 'cors'
-import { QuakeLogRoute } from './routes/quakeLogRoute';
+import { quakeLogRouter } from './routes/quakeLogRouter';
 
 const app = express();
 
 app.use(express.json())
 app.use(cors());
-app.use("/api", QuakeLogRoute);
+app.use("/api", quakeLogRouter);
 
 
 app.listen(3000, () => {
 
-    console.log('api running');
+    console.log('api quake log running');
 });
